Add user interface: buttons, results display
Add buttons for player selections and container to display results on
webpage. Add click event listeners for buttons to run function to use
the clicked button as the player's selection, generate a random choice
for the computer's selection, and run the previously created playRound
function to run the game. Add function to declare results within results
container on webpage and add it to the playRound function.